another fixes
- Fixed Shiren Hat price to make it more consistent with rest of the set
- Moved Old Lady pet and Worms pet codes apart from each other
